Fix timezone issue
The server was running on UTC, meaning that datetime.now() was returning the following day if the time was after 7 or 8pm (depending on DST). Since we only show events starting today onward, this meant that in the evening you'd stop seeing the same day's movies. Fix that by using (an approximation to) the actual EST timezone.